+++ vehicle_options.lua
@@ -2,18 +2,32 @@
 -- Created By Jackz
 -- See changelog in changelog.md
 local SCRIPT = "vehicle_options"
-local VERSION = "1.8.2"
-luahttp = require("luahttp")
-local result = luahttp.request("GET", "jackz.me", "/stand/updatecheck.php?script=" .. SCRIPT .. "&v=" .. VERSION)
-if result == "OUTDATED" then
-    util.toast("A new version of " .. SCRIPT .. " is available")
+local VERSION = "1.12.0"
+-- Remove these lines if you want to disable update-checks: (6-15)
+util.async_http_get("jackz.me", "/stand/updatecheck.php?ucv=2&script=" .. SCRIPT .. "&v=" .. VERSION, function(result)
+    chunks = {}
+    for substring in string.gmatch(result, "%S+") do
+        table.insert(chunks, substring)
+    end
+    if chunks[1] == "OUTDATED" then
+        util.toast(SCRIPT .. " has a new version available.\n" .. VERSION .. " -> " .. chunks[2] .. "\nDownload the latest version from https://jackz.me/sz")
+    end
+end)
+
+local status = pcall(require, "natives-1627063482")
+if not status then
+    util.toast(SCRIPT .. " cannot load: Library files are missing. (natives-1627063482)", 10)
+    util.stop_script()
+end
+local status, json = pcall(require, "json")
+if not status then
+    util.toast(SCRIPT .. " cannot load: Library files are missing. (json)", 10)
+    util.stop_script()
 end
 
--- TODO: Clone Vehicle
--- TODO: Save / Spawn Saved Per-Player
-
-require("natives-1627063482")
-json = require('json')
+--TODO: Idea: Cloud vehicles. Submenu:
+-- "Spawn" "Spawn Inside" "Uplaod to Cloud Vehicles"
+-- Download from Cloud Vehicles
 
 -- Per-player options
 local options = {}
@@ -24,7 +38,7 @@ local DOOR_NAMES = {
     "Engine", "Trunk",
     "Back", "Back 2",
 }
--- Subtract index by 1 to get modIndex (ty lua)
+-- Subtract index by 1 to get modType (ty lua)
 local MOD_TYPES = {
     [1] = "Spoilers",
     [2] = "Front Bumper",
@@ -43,9 +57,6 @@ local MOD_TYPES = {
     [15] = "Horns",
     [16] = "Suspension",
     [17] = "Armor",
-    [19] = "Turbo Turning",
-    [21] = "Tire Smoke",
-    [22] = "Xenon Headlights",
     [24] = "Wheels Design",
     [25] = "Motorcycle Back Wheel Design",
     [26] = "Plate Holders",
@@ -57,6 +68,15 @@ local MOD_TYPES = {
     [36] = "Plaques",
     [39] = "Hydraulics",
     [49] = "Livery"
+}
+-- Subtract index by 1 to get modType (ty lua)
+local TOGGLE_MOD_TYPES = {
+    [18] = "UNK17",
+    [19] = "Turbo Turning",
+    [20] = "UNK19",
+    [21] = "Tire Smoke",
+    [22] = "UNK21",
+    [23] = "Xenon Headlights"
 }
 local VEHICLE_TYPES = {
     "Compacts",  
@@ -82,6 +102,9 @@ local VEHICLE_TYPES = {
 	"Commercial",
 	"Trains"
 }
+local VEHICLE_SAVEDATA_FORMAT_VERSION = "JSTAND 1.1"
+local TOW_TRUCK_MODEL_1 = util.joaat("towtruck")
+local TOW_TRUCK_MODEL_2 = util.joaat("towtruck2")
 local NEON_INDICES = { "Left", "Right", "Front", "Back"}
 
 -- Gets the player's vehicle, attempts to request control. Returns 0 if unable to get control
@@ -143,6 +166,80 @@ function get_player_vehicle_in_control(pid)
     end
     return vehicle
 end
+local CAB_MODEL = util.joaat("phantom")
+local TRAILER_MODEL = util.joaat("tr2")
+function spawn_cab_and_trailer_for_vehicle(vehicle, rampDown) 
+    STREAMING.REQUEST_MODEL(CAB_MODEL)
+    STREAMING.REQUEST_MODEL(TRAILER_MODEL)
+    while not STREAMING.HAS_MODEL_LOADED(TRAILER_MODEL) do
+        util.yield()
+    end
+    ENTITY.SET_ENTITY_VELOCITY(vehicle, 0, 0, 0)
+    local cabPos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, -5.0, 8, 0.0)
+    local trailerPos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, -5.0, 0, 0.0)
+    local heading = ENTITY.GET_ENTITY_HEADING(vehicle)
+
+    local cab = util.create_vehicle(CAB_MODEL, cabPos, heading)
+    local trailer = util.create_vehicle(TRAILER_MODEL, trailerPos, heading)
+    if rampDown then
+        VEHICLE.SET_VEHICLE_DOOR_OPEN(trailer, 5, 0, 0)
+    end
+    VEHICLE.ATTACH_VEHICLE_TO_TRAILER(cab, trailer, 5)
+    local pos = ENTITY.GET_ENTITY_COORDS(vehicle)
+    VEHICLE.ATTACH_VEHICLE_ON_TO_TRAILER(vehicle, trailer, 0, 0, -2.0, 0, 0, 0.0, 0, 0, 0, 0.0)
+    ENTITY.DETACH_ENTITY(vehicle)
+    local driver = PED.CREATE_RANDOM_PED_AS_DRIVER(cab, true)
+    util.yield(1)
+    return cab, driver
+end
+
+local CARGOBOB_MODEL = util.joaat("cargobob")
+function spawn_cargobob_for_vehicle(vehicle) 
+    STREAMING.REQUEST_MODEL(CARGOBOB_MODEL)
+    while not STREAMING.HAS_MODEL_LOADED(CARGOBOB_MODEL) do
+        util.yield()
+    end
+    local pos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, 0, 0, 10.0)
+    local z = memory.alloc(8)
+    if(MISC.GET_GROUND_Z_FOR_3D_COORD(pos.x, pos.y, pos.z, z, false)) then
+        pos.z = memory.read_float(z) + 10.0
+    end
+    memory.free(z)
+    local heading = ENTITY.GET_ENTITY_HEADING(vehicle)
+    VEHICLE.BRING_VEHICLE_TO_HALT(vehicle, 0.0, 10)
+    local cargobob = util.create_vehicle(CARGOBOB_MODEL, pos, heading)
+    VEHICLE._SET_CARGOBOB_HOOK_CAN_DETACH(cargobob, true)
+    VEHICLE._DISABLE_VEHICLE_WORLD_COLLISION(cargobob)
+    VEHICLE.CREATE_PICK_UP_ROPE_FOR_CARGOBOB(cargobob, 0)
+    VEHICLE.SET_HELI_BLADES_FULL_SPEED(cargobob)
+    local driver = PED.CREATE_RANDOM_PED_AS_DRIVER(cargobob, true)
+    ENTITY.SET_ENTITY_VELOCITY(vehicle, 0, 0, 0)
+    ENTITY.SET_ENTITY_VELOCITY(cargobob, 0, 0, 0)
+    VEHICLE.ATTACH_VEHICLE_TO_CARGOBOB(cargobob, vehicle, -2, 0, 0, 0)
+    util.yield(1)
+    return cargobob, driver
+end
+
+function spawn_tow_for_vehicle(vehicle) 
+    local pz = memory.alloc(8)
+    local pos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, 0, 8, 0.1)
+    local heading = ENTITY.GET_ENTITY_HEADING(vehicle)
+    MISC.GET_GROUND_Z_FOR_3D_COORD(pos.x, pos.y, pos.z, pz, true)
+    pos.z = memory.read_float(pz)
+    memory.free(pz)
+    local model = math.random(2) == 2 and TOW_TRUCK_MODEL_1 or TOW_TRUCK_MODEL_2
+    STREAMING.REQUEST_MODEL(model)
+    while not STREAMING.HAS_MODEL_LOADED(model) do
+        util.yield()
+    end
+    local tow = util.create_vehicle(model, pos, heading)
+    ENTITY.SET_ENTITY_VELOCITY(vehicle, 0, 0, 0)
+    VEHICLE.BRING_VEHICLE_TO_HALT(vehicle, 0.0, 5)
+    VEHICLE.ATTACH_VEHICLE_TO_TOW_TRUCK(tow, vehicle, false, 0, 0, 0)
+    local driver = PED.CREATE_RANDOM_PED_AS_DRIVER(tow, true)
+    util.yield(1)
+    return tow, driver, model
+end
 
 
 function setup_action_for(pid) 
@@ -152,6 +249,7 @@ function setup_action_for(pid)
         paint_color_primary = { r = 1.0, g = 0.412, b = 0.706, a = 1 },
         paint_color_secondary = { r = 1.0, g = 0.412, b = 0.706, a = 1 },
         neon_color = { r = 1.0, g = 0.412, b = 0.706, a = 1 },
+        trailer_gate = true
     }
 
     menu.action(submenu, "Teleport Vehicle to Me", {"tpvehme"}, "Teleports their vehicle to your location", function(on_click)
@@ -204,6 +302,187 @@ function setup_action_for(pid)
         end
     end)
 
+    local towMenu = menu.list(movementMenu, "Attachments", {"attachments"}, "Attach their vehicle to:\nCargobobs\nTow Trucks")
+    -- TOW 
+        menu.divider(towMenu, "Tow Trucks")
+        menu.action(towMenu, "Tow (Wander)", {"towwander"}, "Will make a random tow truck tow the player's vehicle randomly around", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle == 0 then
+                util.toast("Player is not in a car or out of range")
+            else
+                local tow, driver = spawn_tow_for_vehicle(vehicle)
+                TASK.TASK_VEHICLE_DRIVE_WANDER(driver, tow, 30.0, 786603)
+            end
+        end)
+
+        menu.action(towMenu, "Tow To Waypoint", {"towwaypoint"}, "Will make a random tow truck tow the player's vehicle to your waypoint", function(on_click)
+            local vehicleModel = ENTITY.GET_ENTITY_MODEL(vehicle)
+            if HUD.IS_WAYPOINT_ACTIVE() then
+                local blip = HUD.GET_FIRST_BLIP_INFO_ID(8)
+                local waypoint_pos = HUD.GET_BLIP_COORDS(blip)
+                local vehicle = get_player_vehicle_in_control(pid)
+                if vehicle == 0 then
+                    util.toast("Player is not in a car or out of range")
+                else
+                    local tow, driver, model = spawn_tow_for_vehicle(vehicle)
+                    TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, tow, waypoint_pos.x, waypoint_pos.y, waypoint_pos.z, 35.0, 1.0, model, 786603, 5.0, 1.0)
+                end
+            else
+                util.toast("You have no waypoint to drive to")
+            end
+        end)
+        BlipColors = {	
+            BLIPCOLOR_NONE = 0x0,	
+            BLIPCOLOR_RED = 0x1,	
+            BLIPCOLOR_GREEN = 0x2,	
+            BLIPCOLOR_BLUE = 0x3,	
+            BLIPCOLOR_PLAYER = 0x4,	
+            BLIPCOLOR_YELLOWMISSION = 0x5,	
+            BLIPCOLOR_FRIENDLYVEHICLE = 0x26,	
+            BLIPCOLOR_MICHAEL = 0x2A,	
+            BLIPCOLOR_FRANKLIN = 0x2B,	
+            BLIPCOLOR_TREAVOR = 0x2C,	
+            BLIPCOLOR_REDMISSION = 0x31,	
+            BLIPCOLOR_MISSIONVEHICLE = 0x36,	
+            BLIPCOLOR_YELLOWMISSION2 = 0x3C,	
+            BLIPCOLOR_MISSION = 0x42,	
+            BLIPCOLOR_WAYPOINT = 0x53
+        }
+
+        function get_player_objective(spriteId)
+            local blip = HUD.GET_FIRST_BLIP_INFO_ID(spriteId)
+            if HUD.DOES_BLIP_EXIST(blip) then
+                local color = HUD.GET_BLIP_COLOUR(blip)
+                if color == BLIPCOLOR_YELLOWMISSION2 or color == BLIPCOLOR_YELLOWMISSION or color == BLIPCOLOR_MISSION or color == BLIPCOLOR_GREEN then
+                    return blip
+                end
+            end
+        end
+
+        menu.action(towMenu, "Detach Tow", {"detachtow"}, "Will detach from any tow truck", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle > 0 then
+                VEHICLE.DETACH_VEHICLE_FROM_ANY_TOW_TRUCK(vehicle)
+            else
+                util.toast("Player is not in a car or out of range")
+            end
+        end)
+
+        menu.divider(towMenu, "Cargobob")
+
+        menu.action(towMenu, "Cargobob To Mt. Chiliad", {"cargobobmt"}, "Will make a random cargobob take the player's vehicle to mount chiliad", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle == 0 then
+                util.toast("Player is not in a car or out of range")
+            else
+                local cargobob, driver = spawn_cargobob_for_vehicle(vehicle)
+                TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, cargobob, 450.718 , 5566.614, 806.183, 100.0, 1.0, CARGOBOB_MODEL, 786603, 5.0, 1.0)
+            end
+        end)
+
+        menu.action(towMenu, "Cargobob To Ocean", {"cargobobocean"}, "Will make a random cargobob take the player's vehicle to the ocean", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle == 0 then
+                util.toast("Player is not in a car or out of range")
+            else
+                local cargobob, driver = spawn_cargobob_for_vehicle(vehicle)
+                local pos = ENTITY.GET_ENTITY_COORDS(cargobob)
+                local vec = memory.alloc(24)
+                local dest = { x = 0, y = 0, z = -5.0}
+                if PATHFIND.GET_NTH_CLOSEST_VEHICLE_NODE(pos.x, pos.y, pos.z, 15, vec, 3, 3.0, 0) then
+                   dest = memory.read_vector3(vec)
+                   memory.free(vec)
+                    dest.z = -5.0
+                else
+                    dest.x = -2156
+                    dest.y = -1311
+                end
+                TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, cargobob, dest.x, dest.y, dest.z, 100.0, 1.0, CARGOBOB_MODEL, 786603, 5.0, 1.0)
+            end
+        end)
+
+        menu.action(towMenu, "Cargobob To Waypoint", {"cargobobwaypoint"}, "Will make a random cargobob take the player's vehicle to your waypoint", function(on_click)
+            local vehicleModel = ENTITY.GET_ENTITY_MODEL(vehicle)
+            if HUD.IS_WAYPOINT_ACTIVE() then
+                local blip = HUD.GET_FIRST_BLIP_INFO_ID(8)
+                local waypoint_pos = HUD.GET_BLIP_COORDS(blip)
+                local vehicle = get_player_vehicle_in_control(pid)
+                if vehicle == 0 then
+                    util.toast("Player is not in a car or out of range")
+                else
+                    local cargobob, driver = spawn_cargobob_for_vehicle(vehicle)
+                    
+                    TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, cargobob, waypoint_pos.x, waypoint_pos.y, waypoint_pos.z, 35.0, 1.0, CARGOBOB_MODEL, 786603, 5.0, 1.0)
+                end
+            else
+                util.toast("You have no waypoint to drive to")
+            end
+        end)
+
+        menu.action(towMenu, "Detach Cargobob", {"detachcargo"}, "Will detach cargobob from vehicle", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle > 0 then
+                VEHICLE.DETACH_VEHICLE_FROM_ANY_CARGOBOB(vehicle)
+            else
+                util.toast("Player is not in a car or out of range")
+            end
+        end)
+
+        menu.divider(towMenu, "Trailers")
+
+        menu.action(towMenu, "Drive Around", {"trailerwander"}, "Will make a random cab & trailer take the player's vehicle randomly around", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle == 0 then
+                util.toast("Player is not in a car or out of range")
+            else
+                local cab, driver = spawn_cab_and_trailer_for_vehicle(vehicle, options[pid].trailer_gate)
+                TASK.TASK_VEHICLE_DRIVE_WANDER(driver, cab, 30.0, 786603)
+            end
+        end)
+
+        menu.action(towMenu, "Take To Waypoint", {"trailerwaypoint"}, "Will make a random cab & trailer take the player's vehicle to your waypoint", function(on_click)
+            local vehicleModel = ENTITY.GET_ENTITY_MODEL(vehicle)
+            if HUD.IS_WAYPOINT_ACTIVE() then
+                local blip = HUD.GET_FIRST_BLIP_INFO_ID(8)
+                local waypoint_pos = HUD.GET_BLIP_COORDS(blip)
+                local vehicle = get_player_vehicle_in_control(pid)
+                if vehicle == 0 then
+                    util.toast("Player is not in a car or out of range")
+                else
+                    local cab, driver = spawn_cab_and_trailer_for_vehicle(vehicle, options[pid].trailer_gate)
+                    TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, cab, waypoint_pos.x, waypoint_pos.y, waypoint_pos.z, 35.0, 1.0, CAB_MODEL, 786603, 5.0, 1.0)
+                end
+            else
+                util.toast("You have no waypoint to drive to")
+            end
+        end)
+
+        menu.toggle(towMenu, "Spawn Trailer With Gate Down", {"trailergate"}, "Should spawned trailers spawn with gate down?", function(on)
+            options[pid].trailer_gate = on
+        end, options[pid].trailer_gate)
+
+        menu.divider(towMenu, "Misc")
+
+        menu.action(towMenu, "Free Vehicle", {"freevehicle"}, "Teleports the vehicle upwards to escape trailers or other objects", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle == 0 then
+                util.toast("Player is not in a car or out of range")
+            else
+                local pos = ENTITY.GET_ENTITY_COORDS(vehicle)
+                ENTITY.SET_ENTITY_COORDS(vehicle, pos.x, pos.y, pos.z + 5.0)
+            end
+        end)
+
+        menu.action(towMenu, "Detach All", {"detachall"}, "Will detach vehicle from anything", function(on_click)
+            local vehicle = get_player_vehicle_in_control(pid)
+            if vehicle > 0 then
+                ENTITY.DETACH_ENTITY(vehicle, true, 0)
+            else
+                util.toast("Player is not in a car or out of range")
+            end
+        end)
+
+    -- END TOW
     menu.action(movementMenu, "Stop", {"stopvehicle"}, "Stops the player's engine", function(on_click)
         local vehicle = get_player_vehicle_in_control(pid)
         if vehicle > 0 then
@@ -341,7 +620,6 @@ function setup_action_for(pid)
     end)
     --END NEON--
 
-    -- TODO: Customize lights (head, rear, neon, etc, toggle)
     menu.action(lsc, "Paint", {"paint"}, "Paints the car with the selected colors below", function(on_click)
         local vehicle = get_player_vehicle_in_control(pid)
         if vehicle > 0 then
@@ -375,10 +653,10 @@ function setup_action_for(pid)
         subMenus = {}
         local vehicle = get_player_vehicle_in_control(pid)
         if vehicle > 0 then
-            -- TODO: Make default value be current value?
             for i, mod in pairs(MOD_TYPES) do
+                local default = VEHICLE.GET_VEHICLE_MOD(vehicle, i -1)
                 local max = VEHICLE.GET_NUM_VEHICLE_MODS(vehicle, i - 1)
-                local m = menu.slider(modMenu, mod, {"set" .. mod}, string.format("Apply upto %d variations of %s", max, mod), 0, max, 0, 1, function(index)
+                local m = menu.slider(modMenu, mod, {"set" .. mod}, string.format("Apply upto %d variations of %s\n-1 is Stock", max, mod), -1, max, default, 1, function(index)
                     local vehicle = get_player_vehicle_in_control(pid)
                     if vehicle > 0 then
                        VEHICLE.SET_VEHICLE_MOD(vehicle, i - 1, index)
@@ -386,6 +664,18 @@ function setup_action_for(pid)
                         util.toast("Player is not in a car or out of range")
                     end
                 end)
+                table.insert(subMenus, m)
+            end
+            for i, mod in pairs(TOGGLE_MOD_TYPES) do
+                local default = VEHICLE.IS_TOGGLE_MOD_ON(vehicle, i-1)
+                local m = menu.toggle(modMenu, mod, {"toggle" .. mod}, "Toggle " .. mod .. " on or off", function(on)
+                    local vehicle = get_player_vehicle_in_control(pid)
+                    if vehicle > 0 then
+                       VEHICLE.TOGGLE_VEHICLE_MOD(vehicle, i-1, on)
+                    else
+                        util.toast("Player is not in a car or out of range")
+                    end
+                end, default)
                 table.insert(subMenus, m)
             end
         end
@@ -504,6 +794,35 @@ function setup_action_for(pid)
 
     end, false)
 
+    menu.action(submenu, "Flip Vehicle 180", {"flipv"}, "Flips the player's vehicle around", function(v)
+        local vehicle = get_player_vehicle_in_control(pid)
+        if vehicle > 0 then
+            local rot = ENTITY.GET_ENTITY_ROTATION(vehicle)
+            ENTITY.SET_ENTITY_ROTATION(vehicle, 0, 0, -rot.z)
+        else
+            util.toast("Player is not in a car or out of range")
+        end
+    end)
+
+    menu.click_slider(submenu, "Hijack Vehicle", {"hijack"}, "Makes a random NPC hijack their vehicle\n 0 = Doors Unlocked\n 1 = Doors Locked", 0, 1, 0, 1, function(hijackLevel)
+        local vehicle = get_player_vehicle_in_control(pid)
+        if vehicle > 0 then
+            local pos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, -2.0, 0.0, 0.1)
+            ENTITY.SET_ENTITY_VELOCITY(vehicle, 0, 0, 0)
+            local ped = PED.CREATE_RANDOM_PED(pos.x, pos.y, pos.z)
+            TASK.TASK_SET_BLOCKING_OF_NON_TEMPORARY_EVENTS(ped, true)
+            PED.SET_BLOCKING_OF_NON_TEMPORARY_EVENTS(ped, true)
+            VEHICLE.SET_VEHICLE_ENGINE_ON(vehicle, true, true)
+            TASK.TASK_VEHICLE_DRIVE_WANDER(ped, vehicle, 100.0, 2883621)
+            if hijackLevel == 1 then
+                util.yield(20)
+                VEHICLE.SET_VEHICLE_DOORS_LOCKED_FOR_ALL_PLAYERS(vehicle, true)
+            end
+        else
+            util.toast("Player is not in a car or out of range")
+        end
+    end)
+
     menu.action(submenu, "Burst Tires", {"bursttires", "bursttyres"}, "Bursts the player's vehicle tires", function(on_click)
         local ped = PLAYER.GET_PLAYER_PED_SCRIPT_INDEX(players.user())
         local pos = ENTITY.GET_ENTITY_COORDS(ped, 1)
@@ -523,10 +842,29 @@ function setup_action_for(pid)
         end
     end)
 
+    menu.action(submenu, "Delete Vehicle", {"deletevehicle"}, "Removes the player's vehicle", function(on_click)
+        local vehicle = get_player_vehicle_in_control(pid)
+        if vehicle > 0 then
+            util.delete_entity(vehicle)
+        else
+            util.toast("Player is not in a car or out of range")
+        end
+    end)
+
     menu.action(submenu, "Kill Engine", {"killengine"}, "Kills the player's vehicle engine", function(on_click)
         local vehicle = get_player_vehicle_in_control(pid)
         if vehicle > 0 then
             VEHICLE.SET_VEHICLE_ENGINE_HEALTH(vehicle, -4000)
+        else
+            util.toast("Player is not in a car or out of range")
+        end
+    end)
+
+    menu.action(submenu, "Clean Vehicle", {"cleanvehicle"}, "Cleans the player's vehicle", function(on_click)
+        local vehicle = get_player_vehicle_in_control(pid)
+        if vehicle > 0 then
+            GRAPHICS.REMOVE_DECALS_FROM_VEHICLE(vehicle)
+            VEHICLE.SET_VEHICLE_DIRT_LEVEL(vehicle, 0.0)
         else
             util.toast("Player is not in a car or out of range")
         end
@@ -654,15 +992,18 @@ function get_vehicle_save_data(vehicle)
     memory.free(Color.g)
     memory.free(Color.b)
 
-    local mods = {}
+    local mods = { Toggles = {} }
     for i, modName in pairs(MOD_TYPES) do
         mods[modName] = VEHICLE.GET_VEHICLE_MOD(vehicle, i - 1)
     end
-
+    for i, mod in pairs(TOGGLE_MOD_TYPES) do
+        mods.Toggles[mod] = VEHICLE.IS_TOGGLE_MOD_ON(vehicle, i-1)
+    end
     local saveData = {
-        Format = "JSTAND 1.0",
+        Format = VEHICLE_SAVEDATA_FORMAT_VERSION,
         Model = model,
-        Name = VEHICLE.GET_DISPLAY_NAME_FROM_VEHICLE_MODEL(model) ,
+        Name = VEHICLE.GET_DISPLAY_NAME_FROM_VEHICLE_MODEL(model),
+        Manufacturer = VEHICLE._GET_MAKE_NAME_FROM_VEHICLE_MODEL(model),
         Type = VEHICLE_TYPES[VEHICLE.GET_VEHICLE_CLASS(vehicle)],
         ["Tire Smoke"] = TireSmoke,
         Livery = {
@@ -686,6 +1027,7 @@ function get_vehicle_save_data(vehicle)
             ["Xenon Color"] = VEHICLE._GET_VEHICLE_XENON_LIGHTS_COLOR(vehicle),
             Neon = Neon
         },
+        ["Engine Running"] = VEHICLE.GET_IS_VEHICLE_ENGINE_RUNNING(vehicle),
         ["Dashboard Color"] = DashboardColor,
         ["Interior Color"] = InteriorColor,
         ["Dirt Level"] = VEHICLE.GET_VEHICLE_DIRT_LEVEL(vehicle),
@@ -696,6 +1038,7 @@ function get_vehicle_save_data(vehicle)
     return saveData
 end
 function apply_vehicle_save_data(vehicle, saveData) 
+    -- Vehicle Paint Colors. Not sure if all these are needed but well I store them
     VEHICLE.SET_VEHICLE_MOD_KIT(vehicle, 0)
     VEHICLE.SET_VEHICLE_COLOUR_COMBINATION(vehicle, saveData.Colors["Color Combo"])
     VEHICLE.SET_VEHICLE_EXTRA_COLOURS(vehicle, saveData.Colors.Extras.r, saveData.Colors.Extras.g, saveData.Colors.Extras.b)
@@ -710,25 +1053,38 @@ function apply_vehicle_save_data(vehicle, saveData)
     else
         VEHICLE.SET_VEHICLE_MOD_COLOR_2(vehicle, saveData.Colors.Secondary["Paint Type"], saveData.Colors.Secondary.Color, saveData.Colors.Secondary["Pearlescent Color"])
     end
+    VEHICLE.SET_VEHICLE_ENVEFF_SCALE(vehicle, saveData["Colors"]["Paint Fade"])
+    -- Misc Colors / Looks
     VEHICLE.SET_VEHICLE_TYRE_SMOKE_COLOR(vehicle, saveData["Tire Smoke"].r, saveData["Tire Smoke"].g, saveData["Tire Smoke"].b)
     VEHICLE.SET_VEHICLE_TYRES_CAN_BURST(vehicle, saveData["Bulletproof Tires"])
     VEHICLE._SET_VEHICLE_DASHBOARD_COLOR(vehicle, saveData["Dashboard Color"])
     VEHICLE._SET_VEHICLE_INTERIOR_COLOR(vehicle, saveData["Interior Color"])
     VEHICLE.SET_VEHICLE_DIRT_LEVEL(vehicle, saveData["Dirt Level"])
-    VEHICLE.SET_VEHICLE_ENVEFF_SCALE(vehicle, saveData["Colors"]["Paint Fade"])
 
-    for i, modName in pairs(MOD_TYPES) do
-        if saveData.Mods[modName] then
-            VEHICLE.SET_VEHICLE_MOD(vehicle, i - 1, saveData.Mods[modName])
-        end
-    end
-
+    -- Lights
+    VEHICLE._SET_VEHICLE_XENON_LIGHTS_COLOR(vehicle, saveData["Lights"]["Xenon Color"])
     VEHICLE._SET_VEHICLE_NEON_LIGHTS_COLOUR(vehicle, saveData.Lights.Neon.Color.r, saveData.Lights.Neon.Color.g, saveData.Lights.Neon.Color.b)
     VEHICLE._SET_VEHICLE_NEON_LIGHT_ENABLED(vehicle, 0, saveData.Lights.Neon.Left)
     VEHICLE._SET_VEHICLE_NEON_LIGHT_ENABLED(vehicle, 1, saveData.Lights.Neon.Right)
     VEHICLE._SET_VEHICLE_NEON_LIGHT_ENABLED(vehicle, 2, saveData.Lights.Neon.Front)
     VEHICLE._SET_VEHICLE_NEON_LIGHT_ENABLED(vehicle, 3, saveData.Lights.Neon.Back)
 
+    if saveData["Engine Running"] then
+        VEHICLE.SET_VEHICLE_ENGINE_ON(vehicle, true, true, false)
+    end
+
+    for i, modName in pairs(MOD_TYPES) do
+        if saveData.Mods[modName] then
+            VEHICLE.SET_VEHICLE_MOD(vehicle, i - 1, saveData.Mods[modName])
+        end
+    end
+    if saveData.Mods.Toggles then
+        for i, mod in pairs(TOGGLE_MOD_TYPES) do
+            VEHICLE.TOGGLE_VEHICLE_MOD(vehicle, i - 1, saveData.Mods.Toggles[mod])
+        end
+    end
+
+    -- Misc
     VEHICLE.SET_VEHICLE_LIVERY(vehicle, saveData.Livery.style)
     VEHICLE.SET_VEHICLE_WINDOW_TINT(vehicle, saveData["Window Tint"])
     VEHICLE.SET_VEHICLE_NUMBER_PLATE_TEXT(vehicle, saveData["License Plate"].Text)
@@ -762,17 +1118,20 @@ menu.on_focus(savedVehiclesList, function()
     savedVehicleMenus = {}
     for _, file in ipairs(filesystem.list_files(vehicleDir)) do
         local _, name, ext = string.match(file, "(.-)([^\\/]-%.?([^%.\\/]*))$")
-            if ext == "json" then
-            local m = menu.action(savedVehiclesList, name, {"spawnvehicle" .. name}, "Spawns a saved custom vehicle", function(on_click)
-                local file = io.open(vehicleDir .. name, "r")
-                io.input(file)
-                local saveData = json.decode(io.read("*a"))
-                io.close(file)
+        if ext == "json" then
+            local file = io.open(vehicleDir .. name, "r")
+            io.input(file)
+            local saveData = json.decode(io.read("*a"))
+            io.close(file)
+            local manuf = saveData.Manufacturer and saveData.Manufacturer .. " " or ""
+            local desc = string.format("Vehicle: %s%s (%s)\nFormat Version: %s", manuf, saveData.Name, saveData.Type, saveData.Format)
+            local displayName = string.sub(name, 0, -6)
+            local m = menu.action(savedVehiclesList, "Spawn", {"spawnvehicle" .. displayName}, "Spawns a saved custom vehicle\n" .. desc, function(on_click)
                 if saveData['Model'] ~= nil then
                     local my_ped = PLAYER.GET_PLAYER_PED_SCRIPT_INDEX(players.user())
                     local pos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(my_ped, 0.0, 5.0, 0.5)
                     local heading = ENTITY.GET_ENTITY_HEADING(target_ped)
-                    local vehicle
+                    local vehicle = 0
                     if applySaved then
                         local v = util.get_vehicle()
                         if v == 0 then
@@ -782,7 +1141,11 @@ menu.on_focus(savedVehiclesList, function()
                         end
                         util.toast("Applied " .. name .. " to your current vehicle")
                     else
-                        util.create_vehicle(saveData.Model, pos, heading)
+                        STREAMING.REQUEST_MODEL(saveData.Model)
+                        while not STREAMING.HAS_MODEL_LOADED(saveData.Model) do
+                            util.yield()
+                        end
+                        vehicle = util.create_vehicle(saveData.Model, pos, heading)
                         util.toast("Spawned " .. name)
                     end
 
@@ -797,7 +1160,129 @@ menu.on_focus(savedVehiclesList, function()
         end
     end
 end)
+local spawned_tows = {}
+menu.action(menu.my_root(), "Tow All Nearby Vehicles", {}, "WARNING: This will make your FPS drop or even crash your game!\nLimited to 30 tow trucks to prevent crashes.", function(sdfa)
+    local pz = memory.alloc(8)
+    STREAMING.REQUEST_MODEL(TOW_TRUCK_MODEL_1)
+    STREAMING.REQUEST_MODEL(TOW_TRUCK_MODEL_2)
+    while not STREAMING.HAS_MODEL_LOADED(TOW_TRUCK_MODEL_2) do
+        util.yield()
+    end
+    for _, ent in ipairs(spawned_tows) do
+        util.delete_entity(ent)
+    end
+    spawned_tows = {}
+    local my_ped = PLAYER.GET_PLAYER_PED_SCRIPT_INDEX(players.user())
+    local my_pos = ENTITY.GET_ENTITY_COORDS(my_ped)
+    local nearby_vehicles = {}
+    for _, vehicle in ipairs(util.get_all_vehicles()) do
+        local model = ENTITY.GET_ENTITY_MODEL(vehicle)
+        if model ~= towtruck_1 and model ~= towtruck_2 and VEHICLE.IS_THIS_MODEL_A_CAR(model) then
+            local pos = ENTITY.GET_OFFSET_FROM_ENTITY_IN_WORLD_COORDS(vehicle, 0, 8, 0.1)
+            MISC.GET_GROUND_Z_FOR_3D_COORD(pos.x, pos.y, pos.z, pz, true)
+            pos.z = memory.read_float(pz)
+            local dist = SYSTEM.VDIST2(my_pos.x, my_pos.y, my_pos.z, pos.x, pos.y, pos.z)
+            table.insert(nearby_vehicles, { vehicle, dist, pos, model })
+        end
+    end
+    memory.free(pz)
+    table.sort(nearby_vehicles, function(a, b) return a[2] > b[2] end)
+    for i = 1,30 do
+        if nearby_vehicles[i] then
+            local vehicle = nearby_vehicles[i][1]
+            local pos = nearby_vehicles[i][3]
+            local heading = ENTITY.GET_ENTITY_HEADING(vehicle)
+            
+            math.randomseed(nearby_vehicles[i][4])
+            local tow = util.create_vehicle(math.random(2) == 2 and TOW_TRUCK_MODEL_1 or TOW_TRUCK_MODEL_2, pos, heading)
+            ENTITY.SET_ENTITY_VELOCITY(vehicle, 0, 0, 0)
+            VEHICLE.ATTACH_VEHICLE_TO_TOW_TRUCK(tow, vehicle, false, 0, 0, 0)
+            local driver = PED.CREATE_RANDOM_PED_AS_DRIVER(tow, true)
+            util.yield(1)
+            TASK.TASK_VEHICLE_DRIVE_WANDER(driver, tow, 30.0, 786603)
+            table.insert(spawned_tows, tow)
+            table.insert(spawned_tows, driver)
+        end
+    end
+end)
+menu.action(menu.my_root(), "Clear All Nearby Tows", {}, "", function(sdfa)
+    local vehicles = util.get_all_vehicles()
+    for _, vehicle in ipairs(vehicles) do
+        local model = ENTITY.GET_ENTITY_MODEL(vehicle)
+        if model == TOW_TRUCK_MODEL_1 or model == TOW_TRUCK_MODEL_2 then
+            local driver = VEHICLE.GET_PED_IN_VEHICLE_SEAT(vehicle, -1)
+            if driver > 0 and not PED.IS_PED_A_PLAYER(driver) then
+                util.delete_entity(driver)
+            end
+            util.delete_entity(vehicle)
+        end
+    end
+end)
+menu.action(menu.my_root(), "Cargobob Nearby Cars", {}, "Ignores players, so they can watch :)", function(a)
+    local ped = PLAYER.GET_PLAYER_PED_SCRIPT_INDEX(players.user())
+    local pos = ENTITY.GET_ENTITY_COORDS(ped, 1)
 
+    local cargobobs = {}
+    for _, vehicle in ipairs(util.get_all_vehicles()) do
+        local model = ENTITY.GET_ENTITY_MODEL(vehicle)
+        if VEHICLE.IS_THIS_MODEL_A_CAR(model) and not ENTITY.IS_ENTITY_ATTACHED_TO_ANY_VEHICLE(vehicle) then
+            local driver = VEHICLE.GET_PED_IN_VEHICLE_SEAT(vehicle, -1)
+            if driver == 0 or not PED.IS_PED_A_PLAYER(driver) then
+                local pos2 = ENTITY.GET_ENTITY_COORDS(vehicle, 1)
+                local dist = SYSTEM.VDIST(pos.x, pos.y, pos.z, pos2.x, pos2.y, pos2.z)
+                if dist <= 100.0 then
+                    local cargobob, driver = spawn_cargobob_for_vehicle(vehicle)
+                    VEHICLE._SET_CARGOBOB_HOOK_CAN_DETACH(cargobob, false)
+                    VEHICLE._DISABLE_VEHICLE_WORLD_COLLISION(cargobob)
+                    ENTITY.SET_ENTITY_COLLISION(cargobob, false, false)
+                    ENTITY.SET_ENTITY_INVINCIBLE(cargobob, true)
+                    table.insert(cargobobs, cargobob)
+                    TASK.TASK_VEHICLE_DRIVE_TO_COORD(driver, cargobob, 450.718 , 5566.614, 806.183, 100.0, 1.0, CARGOBOB_MODEL, 786603, 5.0, 1.0)
+                end
+            end
+        end
+    end
+    util.yield(1000)
+    util.create_tick_handler(function(a)
+        local iterations = 0
+        while true do
+            for i, cargo in ipairs(cargobobs) do
+                local v = VEHICLE.GET_VEHICLE_ATTACHED_TO_CARGOBOB(cargo)
+                if v == 0 or VEHICLE.GET_HELI_MAIN_ROTOR_HEALTH(cargo) <= 0 or VEHICLE.GET_HELI_TAIL_ROTOR_HEALTH(cargo) < 0 or VEHICLE.GET_VEHICLE_ENGINE_HEALTH(cargo) <= 0 then
+                    local driver = VEHICLE.GET_PED_IN_VEHICLE_SEAT(cargo, -1)
+                    ENTITY.DETACH_ENTITY(cargo)
+                    util.delete_entity(driver)
+                    util.delete_entity(cargo)
+                    table.remove(cargobobs, i)
+                end
+                util.yield(2000)
+                iterations = iterations + 1
+                if iterations >= 15 then
+                    break
+                end
+            end
+        end
+        util.yield(3000)
+        for _, cargo in ipairs(cargobobs) do
+            ENTITY.SET_ENTITY_INVINCIBLE(cargo, false)
+            VEHICLE._SET_CARGOBOB_HOOK_CAN_DETACH(cargobob, true)
+        end
+        return false
+    end)
+end)
+menu.action(menu.my_root(), "Clear All Nearby Cargobobs", {}, "", function(sdfa)
+    local vehicles = util.get_all_vehicles()
+    for _, vehicle in ipairs(vehicles) do
+        local model = ENTITY.GET_ENTITY_MODEL(vehicle)
+        if model == CARGOBOB_MODEL then
+            local driver = VEHICLE.GET_PED_IN_VEHICLE_SEAT(vehicle, -1)
+            if driver > 0 and not PED.IS_PED_A_PLAYER(driver) then
+                util.delete_entity(driver)
+            end
+            util.delete_entity(vehicle)
+        end
+    end
+end)
 
 local cur_players = players.list(true, true, true)
 for k,pid in pairs(cur_players) do